backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
supabase
python-dotenv
